🛡️ Sentinel: [CRITICAL] Fix integer overflow in memory allocations

+++ src/fastlp.c
@@ -29,9 +29,9 @@ void fastlp(double *obj, double *mat, double *rhs, int *m0 , int *n0, double *op
     fc_solver_state_t *state = fc_solver_state_create(m, n, nz);
     
     // Arrays strictly built by driver:
-    int *ia = (int*)FC_CALLOC(nz + m, sizeof(int));
-    int *ka = (int*)FC_CALLOC(n + m + 1, sizeof(int));
-    double *a = (double*)FC_CALLOC(nz + m, sizeof(double));
+    int *ia = (int*)FC_CALLOC((size_t)nz + m, sizeof(int));
+    int *ka = (int*)FC_CALLOC((size_t)n + m + 1, sizeof(int));
+    double *a = (double*)FC_CALLOC((size_t)nz + m, sizeof(double));
 
     k = 0;
     for (j=0; j<n; j++) {
